feat: replace HorizontalFlowingCards with FeaturedIssues component on homepage and implement optimized card display
