+++ client/src/app/page.tsx
@@ -2,14 +2,14 @@
 
 import Navbar from "@/components/navbar-components/navbar";
 import HeroHeader from "@/components/home/hero-header";
-import Demo from "@/components/ui/horizontal-flowing-cards";
+import FeaturedIssues from "@/components/home/featuerd-issues";
 
 export default function Home() {
   return (
     <>
       <Navbar />
       <HeroHeader />
-      <Demo />
+      <FeaturedIssues />
     </>
   );
 }
