+++ client/src/components/home/featuerd-issues/data.ts
@@ -1,0 +1,49 @@
+import type { CardData } from "./types";
+
+export const featuredIssues: CardData[] = [
+  {
+    id: 1,
+    title: "UNEXPECTED AI BEAUTY",
+    subtitle: "Exploring the Harmony Between Humanity and Nature Through Digital Art",
+    category: "Tech & Art",
+    image: "https://images.unsplash.com/photo-1634017839464-5c339ebe3cb4?w=600&h=800&fit=crop&auto=format&q=80",
+    date: "2024-10-01",
+    pages: 105
+  },
+  {
+    id: 2,
+    title: "ENCODED COUTURE",
+    subtitle: "Aesthetic Intelligence Meets Machine Made Fashion",
+    category: "Fashion",
+    image: "https://images.unsplash.com/photo-1469334031218-e382a71b716b?w=600&h=800&fit=crop&auto=format&q=80",
+    date: "2024-12-01",
+    pages: 105
+  },
+  {
+    id: 3,
+    title: "EMOTIVE MACHINES",
+    subtitle: "DNA of Intelligent Empathy in Modern Technology",
+    category: "Psychology",
+    image: "https://images.unsplash.com/photo-1518709268805-4e9042af2176?w=600&h=800&fit=crop&auto=format&q=80",
+    date: "2024-12-01",
+    pages: 105
+  },
+  {
+    id: 4,
+    title: "DIGITAL DREAMS",
+    subtitle: "Virtual Reality and the Future of Human Experience",
+    category: "VR/AR",
+    image: "https://images.unsplash.com/photo-1592478411213-6153e4ebc696?w=600&h=800&fit=crop&auto=format&q=80",
+    date: "2024-11-15",
+    pages: 98
+  },
+  {
+    id: 5,
+    title: "QUANTUM LEAP",
+    subtitle: "Breaking Barriers in Computational Science",
+    category: "Science",
+    image: "https://images.unsplash.com/photo-1635070041078-e363dbe005cb?w=600&h=800&fit=crop&auto=format&q=80",
+    date: "2025-01-20",
+    pages: 120
+  }
+];
